consolidate physical_optimizer tests into core/tests/physical_optimizer (#14244)
* consolidate physical_optimizer tests into core/tests/physical_optimizer

Move tests

* cleanu
