+++ datafusion/physical-optimizer/src/limited_distinct_aggregation.rs
@@ -190,4 +190,4 @@ impl PhysicalOptimizerRule for LimitedDistinctAggregation {
     }
 }
 
-// See tests in datafusion/core/tests/physical_optimizer/limited_distinct_aggregation.rs
+// See tests in datafusion/core/tests/physical_optimizer
